Center date range in datepicker. Update reduction level selection to use computed property.
Mapping the reduction level selection to a computed property allows us to do the advanced logic we need to properly filter. When a reduction level is selected, the setter will translate that reduction level into the appropriate query parameters. The getter is used to set the proper dropdown text based on the currently set query params.