Remove call to aggregate endpoint upon successful data retrieval.
We want to call the aggregate endpoint with new params when 1) The page is initially loaded, 2) When a filter is updated and data refreshed, 3) When a new time range is selected. This matches the previous behavior of the archive client and prevents a slew of requests to the API aggregate endpoint every time a user pages over in the frontend.